# Bytes-only interface