--- a/bindings/csharp/ckzg.c
+++ b/bindings/csharp/ckzg.c
@@ -24,55 +24,25 @@ void free_trusted_setup_wrap(KZGSettings *s) {
   free(s);
 }
 
-C_KZG_RET blob_to_kzg_commitment_wrap(uint8_t out[48], const Blob blob, const KZGSettings *s) {
-  KZGCommitment c;
-  C_KZG_RET ret;
-  ret = blob_to_kzg_commitment(&c, blob, s);
-  if (ret != C_KZG_OK) return ret;
-  bytes_from_g1(out, &c);
-  return C_KZG_OK;
+C_KZG_RET blob_to_kzg_commitment_wrap(KZGCommitment *commitment, const Blob blob, const KZGSettings *s) {
+  return blob_to_kzg_commitment(commitment, blob, s);
 }
 
-int verify_aggregate_kzg_proof_wrap(const Blob blobs[], const uint8_t commitments[], size_t n, const uint8_t proof[48], const KZGSettings *s) {
-  KZGProof f;
-  C_KZG_RET ret;
-  ret = bytes_to_g1(&f, proof);
-  if (ret != C_KZG_OK) return -1;
-
-  KZGCommitment* c = calloc(n, sizeof(KZGCommitment));
-  if (c == NULL) return -2;
-
-  for (size_t i = 0; i < n; i++) {
-    ret = bytes_to_g1(&c[i], &commitments[i * 48]);
-    if (ret != C_KZG_OK) { free(c); return -1; }
-  }
-
+int verify_aggregate_kzg_proof_wrap(const Blob blobs[], const KZGCommitment *commitments, size_t n, const KZGProof *proof, const KZGSettings *s) {
   bool b;
-  ret = verify_aggregate_kzg_proof(&b, blobs, c, n, &f, s);
-  free(c);
+  C_KZG_RET ret = verify_aggregate_kzg_proof(&b, blobs, commitments, n, proof, s);
   if (ret != C_KZG_OK) return -1;
 
   return b ? 0 : 1;
 }
 
-C_KZG_RET compute_aggregate_kzg_proof_wrap(uint8_t out[48], const Blob blobs[], size_t n, const KZGSettings *s) {
-  KZGProof f;
-  C_KZG_RET ret;
-  ret = compute_aggregate_kzg_proof(&f, blobs, n, s);
-  if (ret != C_KZG_OK) return ret;
-  bytes_from_g1(out, &f);
-  return C_KZG_OK;
+C_KZG_RET compute_aggregate_kzg_proof_wrap(KZGProof *out, const Blob blobs[], size_t n, const KZGSettings *s) {
+  return compute_aggregate_kzg_proof(out, blobs, n, s);
 }
 
-int verify_kzg_proof_wrap(const uint8_t c[48], const uint8_t z[32], const uint8_t y[32], const uint8_t p[48], KZGSettings *s) {
-  KZGCommitment commitment;
-  KZGProof proof;
+int verify_kzg_proof_wrap(const KZGCommitment *c, const BLSFieldElement *z, const BLSFieldElement *y, const KZGProof *p, KZGSettings *s) {
   bool out;
-
-  if (bytes_to_g1(&commitment, c) != C_KZG_OK) return -1;
-  if (bytes_to_g1(&proof, p) != C_KZG_OK) return -1;
-
-  if (verify_kzg_proof(&out, &commitment, z, y, &proof, s) != C_KZG_OK)
+  if (verify_kzg_proof(&out, c, z, y, p, s) != C_KZG_OK)
     return -2;
 
   return out ? 0 : 1;