--- a/bindings/csharp/ckzg.h
+++ b/bindings/csharp/ckzg.h
@@ -13,10 +13,10 @@ DLLEXPORT KZGSettings* load_trusted_setup_wrap(const char* file);
 
 DLLEXPORT void free_trusted_setup_wrap(KZGSettings *s);
 
-DLLEXPORT C_KZG_RET blob_to_kzg_commitment_wrap(KZGCommitment *out, const Blob blob, const KZGSettings *s);
+DLLEXPORT C_KZG_RET blob_to_kzg_commitment(KZGCommitment *out, const Blob blob, const KZGSettings *s);
 
 DLLEXPORT int verify_aggregate_kzg_proof_wrap(const Blob blobs[], const KZGCommitment *commitments, size_t n, const KZGProof *proof, const KZGSettings *s);
 
-DLLEXPORT C_KZG_RET compute_aggregate_kzg_proof_wrap(KZGProof *out, const Blob blobs[], size_t n, const KZGSettings *s);
+DLLEXPORT C_KZG_RET compute_aggregate_kzg_proof(KZGProof *out, const Blob blobs[], size_t n, const KZGSettings *s);
 
 DLLEXPORT int verify_kzg_proof_wrap(const KZGCommitment *c, const BLSFieldElement *z, const BLSFieldElement *y, const KZGProof *p, KZGSettings *s);