--- a/bindings/java/c_kzg_4844_jni.c
+++ b/bindings/java/c_kzg_4844_jni.c
@@ -144,7 +144,7 @@ JNIEXPORT jbyteArray JNICALL Java_ethereum_ckzg4844_CKZG4844JNI_computeAggregate
   size_t count_native = (size_t)count;
   jbyte *blobs_native = (*env)->GetByteArrayElements(env, blobs, NULL);
   jbyteArray proof = (*env)->NewByteArray(env, BYTES_PER_PROOF);
-  KZGProof *proof_native = (KZGProof *)(uint8_t *)(*env)->GetByteArrayElements(env, proofBytes, NULL);
+  KZGProof *proof_native = (KZGProof *)(uint8_t *)(*env)->GetByteArrayElements(env, proof, NULL);
 
   C_KZG_RET ret = compute_aggregate_kzg_proof(proof_native, (uint8_t const(*)[BYTES_PER_BLOB])blobs_native, count_native, settings);
 
@@ -184,7 +184,7 @@ JNIEXPORT jboolean JNICALL Java_ethereum_ckzg4844_CKZG4844JNI_verifyAggregateKzg
   bool out;
   C_KZG_RET ret = verify_aggregate_kzg_proof(&out, (uint8_t const(*)[BYTES_PER_BLOB])blobs_native, commitments_native, count_native, proof_native, settings);
 
-  (*env)->ReleaseByteArrayElements(env, proof, proof_native, JNI_ABORT);
+  (*env)->ReleaseByteArrayElements(env, proof, (jbyte *)proof_native, JNI_ABORT);
   (*env)->ReleaseByteArrayElements(env, commitments, (jbyte *)commitments_native, JNI_ABORT);
   (*env)->ReleaseByteArrayElements(env, blobs, blobs_native, JNI_ABORT);
 
@@ -217,7 +217,7 @@ JNIEXPORT jbyteArray JNICALL Java_ethereum_ckzg4844_CKZG4844JNI_blobToKzgCommitm
   KZGCommitment *commitment_native = (KZGCommitment *)(*env)->GetByteArrayElements(env, commitment, NULL);
 
 
-  C_KZG_RET ret = blob_to_kzg_commitment(&c, (uint8_t *)blob_native, settings);
+  C_KZG_RET ret = blob_to_kzg_commitment(commitment_native, (uint8_t *)blob_native, settings);
 
   (*env)->ReleaseByteArrayElements(env, blob, blob_native, JNI_ABORT);
   (*env)->ReleaseByteArrayElements(env, commitment, (jbyte *)commitment_native, JNI_ABORT);
@@ -245,12 +245,12 @@ JNIEXPORT jboolean JNICALL Java_ethereum_ckzg4844_CKZG4844JNI_verifyKzgProof(JNI
   BLSFieldElement *y_native = (BLSFieldElement *)(*env)->GetByteArrayElements(env, y, NULL);
 
   bool out;
-  ret = verify_kzg_proof(&out, &c, (uint8_t *)z_native, (uint8_t *)y_native, &p, settings);
+  C_KZG_RET ret = verify_kzg_proof(&out, commitment_native, z_native, y_native, proof_native, settings);
 
-  (*env)->ReleaseByteArrayElements(env, commitment, commitment_native, JNI_ABORT);
-  (*env)->ReleaseByteArrayElements(env, z, z_native, JNI_ABORT);
-  (*env)->ReleaseByteArrayElements(env, y, y_native, JNI_ABORT);
-  (*env)->ReleaseByteArrayElements(env, proof, proof_native, JNI_ABORT);
+  (*env)->ReleaseByteArrayElements(env, commitment, (jbyte *)commitment_native, JNI_ABORT);
+  (*env)->ReleaseByteArrayElements(env, z, (jbyte *)z_native, JNI_ABORT);
+  (*env)->ReleaseByteArrayElements(env, y, (jbyte *)y_native, JNI_ABORT);
+  (*env)->ReleaseByteArrayElements(env, proof, (jbyte *)proof_native, JNI_ABORT);
 
   if (ret != C_KZG_OK)
   {