This PR is an alternative to #37. See the notable changes below:

For C code:

* Replace `KZGCommitment`, `KZGProof`, and `BLSFieldElement` with bytes-only struct.
* In exported C funcs, add conversions to/from `g1_t`/`fr_t` point types.
* Remove exported conversions functions (like `bytes_to_g1`).
  * Make those functions static.
* Change type for field elements from `uint8_t z[BYTES_PER_FIELD_ELEMENT]` to `BLSFieldElement *z`.
  * Easier to read/understand IMO.
* Where applicable, use pointer style instead of array brackets.
  * For example, `void f(char a[])` -> `void f(char *a)`.
  * Not a strong reason for this besides consistency.
* For `verify_aggregate_kzg_proof`, use singular exit point during allocations.
* Wrap `BYTES_PER_BLOB` value in parentheses.
  * Ran into an issue with python bindings, where `n = 393216` (3 blobs).
  * `n = n / BYTES_PER_BLOB` would result the incorrect value: 3072.
  * This is because it was evaluated like `(n/32)*4096` instead of `n/(32*4096)`.

For Java bindings:

* Remove calls to `bytes_from_g1` and `bytes_to_g1`.
  * Also got to remove a bunch of allocations associated with that.
* Add `.PHONY` tags to rules so these get executed every time.
  * Ran into a cached build issue that misled me into thinking things worked.

For C# bindings:

* Remove calls to `bytes_from_g1` and `bytes_to_g1`.
* Remove `blob_to_kzg_commitment_wrap` and `compute_aggregate_kzg_proof_wrap` functions.
  * These are identical to their respective functions now.
* Remove `returns 0 on success` comment. Dup of docstring comment right above it.
* Fix some real minor spacing nits in the tests.

For Node.js bindings:

* Remove calls to `bytes_from_g1` and `bytes_to_g1`.
* Update expected exception in test.
  * It would expect a failure in `bytes_to_g1` which got removed.

For Python bindings:

* Change parameter/return types from `PyCapsule` to `PyBytes`.
  * The capsule bytes cannot be accessed which makes it difficult to actually use.
* Flatten `commitments` in `verify_aggregate_kzg_proof_wrap`.
  * Don't need to iterate of the sequence & extract anymore.
* Return value is allocated with `PyBytes_FromStringAndSize` now.
  * This is faster and simpler.
  * Can remove the `free_G1` method now.
* Remove `bytes_from_g1_wrap` as it's not needed anymore.
* Specify Python version in Makefile.
  * Spent about 5 minutes trying to figure out why python development headers couldn't be found.
  * On my system (and probably others), `python` doesn't point to anything.