@asn-d6 2023-01-12
In what way is this PR an alternative to #37? The idea with #37 is that client devs would never need to interact with any cryptographic objects and they would just pass bytes to the library. Does this PR have the same goals?