@jtraglia 2023-01-12
Also, this is ready for review now. A few things to note:

* The python bindings compile, but I was unable to run the tests on my development system. Something about an unknown option `-Bsymbolic` during linking. I was uncertain about the change in [py_ecc_tests.py](https://github.com/ethereum/c-kzg-4844/pull/62/files#diff-565a335cf3223b4d9fec8fe48c188a0abdfb53928769fa5222004dd8bbdff84eL64-R64).
* I'm not confident that the C# tests are properly testing. I made some bad changes, ran the test, and it seems fine. It could be a cached build or something, not sure. Worth looking into. Also, if it were up to me, I'd remove the other two wrapped functions (`verify_aggregate_kzg_proof_wrap` and `verify_kzg_proof_wrap`) but I see the appeal to having `ret=0` be the success value for all of the functions.