@jtraglia 2023-01-12
Regarding performance, benchmarks are just a hair slower but that makes sense because the conversion from G1 to bytes was done in setup and wasn't including in the benchmark time. In reality, it's probably the exact same overall.

Before:
```
$ ./bench.sh 
goos: darwin
goarch: arm64
pkg: github.com/jtraglia/cgo-kzg-4844
BenchmarkBytesFromG1-20                         463127452               78.02 ns/op          192 B/op          2 allocs/op
BenchmarkBytesToG1-20                           89462080               402.3 ns/op           192 B/op          2 allocs/op
BenchmarkBytesToBlsField-20                     87647565               411.7 ns/op            64 B/op          2 allocs/op
BenchmarkComputeAggregateKzgProof-20                 730          49276212 ns/op             336 B/op          3 allocs/op
BenchmarkVerifyAggregateKzgProof-20                15478           2330489 ns/op             145 B/op          2 allocs/op
BenchmarkBlobToKzgCommitment-20                    18087           1990456 ns/op          131408 B/op          4 allocs/op
BenchmarkVerifyKzgProof-20                         41899            858431 ns/op             353 B/op          5 allocs/op
```

After:
```
$ ./bench.sh
goos: darwin
goarch: arm64
pkg: github.com/jtraglia/cgo-kzg-4844
BenchmarkComputeAggregateKzgProof-20                 721          49401536 ns/op              48 B/op          1 allocs/op
BenchmarkVerifyAggregateKzgProof-20                15142           2381334 ns/op              49 B/op          2 allocs/op
BenchmarkBlobToKzgCommitment-20                    17972           2009970 ns/op          131120 B/op          2 allocs/op
BenchmarkVerifyKzgProof-20                         40689            891162 ns/op             161 B/op          5 allocs/op
```

(note the removed functions are no longer benchmarked)