@jtraglia 2023-01-13
> WRT "this is ready for review", there were no modifications required for the Rust bindings?

Hey George, this is associated with some chats in telegram. After marking it as ready to review, it was decided to merge the rust bindings (#4) into main. Then main got merged into this PR's branch. @pawanjay176 (who wrote the rust bindings) said that he will update them to the bytes interface today.